+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 # 3.0.1
 * Fixated OIDC Server Mock version to latest minor (`1.2`) to prevent future random test failures
+* Updated demo / integration tests framework
 
 # 3.0.0
 * Updated to Spring Boot 4.x